changed lat/long to floats

+++ hw2/localee/db/migrate/20120924045423_create_locations.rb
@@ -2,8 +2,8 @@ class CreateLocations < ActiveRecord::Migration
   def change
     create_table :locations do |t|
       t.string :name
-      t.double :latitude
-      t.double :longitude
+      t.float :latitude
+      t.float :longitude
 
       t.timestamps
     end
